+++ R/lockfile.R
@@ -174,17 +174,27 @@ renv_lockfile_create <- function(project,
 
   lockfile <- renv_lockfile_init(project)
 
-  renv_lockfile_records(lockfile) <-
+  repeat {
+    withRestarts(
+      {
+        records <- renv_snapshot_libpaths(libpaths = libpaths, project = project)
+        records <- renv_snapshot_filter(
+          project  = project,
+          records  = records,
+          type     = type,
+          packages = packages,
+          exclude  = exclude
+        )
+        break
+      },
+      recomputeRecords = function() {
+        renv_dynamic_reset()
+      }
+    )
+  }
 
-    renv_snapshot_libpaths(libpaths = libpaths,
-                           project  = project) %>%
-
-    renv_snapshot_filter(project  = project,
-                         type     = type,
-                         packages = packages,
-                         exclude  = exclude) %>%
-
-    renv_snapshot_fixup()
+  records <- renv_snapshot_fixup(records)
+  renv_lockfile_records(lockfile) <- records
 
   lockfile <- renv_lockfile_fini(lockfile, project)